Crate rusty_picture_namer

Functions

alt_get_filetypes

Uses a default list of filetypes. The default list is read in as a String to keep this function similar to the main get_filetypes function.

arg_parser_2

Parses the arg for the path to the directory containing the files to be renamed. Uses a bundled list of filetypes to provide the filetypes used to idenitfy pictures.

arg_parser_3

Parses two inputted args where the first one is the path to the directory with the files to be renamed and the second one is the path to a list containing filetypes. This supports additional file formats.

directory_walker

Walks the directories to ensure that all pictures get renamed. If there are pictures in subdirectories they will get renamed.

file_counter

Counts the files to be renamed. Some files may already have the directory name already prepended so no rename needs to be done.

file_namer

This renames the files with the specified filetypes.

get_filetypes

Reads a text file into a String and parses it into a vector containing the filetypes.

lead_zeros

Checks to make sure that a situation where the length of the string with leading zeros can support the amount of files in the directory.

log_writer

Writes the contents of String: log_contents into a “.log” file with the name: rusty_picture_namer_YYYY-MM-DD_HHMMSS.log at location: folder_path.

main

One arg <folder_path> which provides the path to the files that need to be renamed is required. The other arg <list_of_filetyps> is optional. If used it will provide an alternate list of filetypes to use.

modified_duration

Returns how long ago a file was modified. A time to compare has to be provided to ensure that all comparisions are compared to the same time.

vec_contains

This is used since option_result_contains for vectors is unstable. This checks is a vector made of Strings contains a string.

zfill

Creates and returns a String of length new_length with leading zeros.